+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 - migrated from Yarn to PNPM (Thorsten)
 - migrated from Fork Awesome to Bootstrap Icons (Thorsten, Jan Harms)
 - migrated codebase to use PHP 8.2 language features (Thorsten)
+- migrated Markdown class (Thorsten)
 - updated to PHPUnit v11 (Thorsten)
 - updated Polish translation (Zięba Bogusław Chaffinch)